# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spartan Cloud APIs",
    "version": "1.0.0",
    "description": "Powered by Costal Cloud. Current master-to-node provisioning and standalone routing APIs, plus optional provisioning and recovery Worker endpoints. HMAC-SHA256 signs timestamp + newline + uppercase method + newline + pathname + newline + raw body. Timestamps are milliseconds within five minutes; GET signs an empty body. Credentials are literal strings. Exact raw bytes must be sent; do not follow redirects. Header auth controls cannot calculate signatures automatically. See API.md for full integration flow and current regional-routing limitations."
  },
  "servers": [
    {
      "url": "https://{worker}.{account}.workers.dev",
      "description": "Default routing-only Worker",
      "variables": {
        "worker": {
          "default": "dezerx-spartant-cloud"
        },
        "account": {
          "default": "YOUR_ACCOUNT_SUBDOMAIN"
        }
      }
    }
  ],
...
  "paths": {
...
    "/control/provision": {
      "post": {
        "operationId": "provisionNodeTenant",
        "summary": "Create or resume tenant container on one node",
        "tags": [
          "Nodes"
        ],
        "description": "Call both nodes with the same persistent ID, key, fingerprint, primary and lifecycle version. Send initialAdmin only to the primary. A 200 provisioning response requires retries. Maximum body 16,384 bytes. Node operational and most validation failures return 503 operation_failed.",
        "responses": {
          "200": {
            "description": "Inspect the response body state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeProvisionResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON; most other node validation failures return 503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or timestamp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource or route not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Node validation or operational failure: operation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "Timestamp": [],
            "Signature": []
          }
        ],
        "x-hmac-secret": "NODE_CONTROL_SECRET",
        "servers": [
          {
            "url": "https://node-us.yourdomain.com",
            "description": "US node; replace hostname"
          },
          {
            "url": "https://node-de.yourdomain.com",
            "description": "Germany node; replace hostname"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeProvisionInput"
              },
              "example": {
                "id": "t-111111111111111111111111",
                "primary": "us",
                "appKey": "base64:AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=",
                "url": "https://t-111111111111111111111111.cloud.yourdomain.com",
                "fingerprint": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                "lifecycleVersion": 0,
                "initialAdmin": {
                  "displayName": "Customer Owner",
                  "email": "owner@customer.example",
                  "password": "Replace-this-password!"
                }
              }
            }
          }
        }
      }
    },
    "/control/lifecycle": {
      "post": {
        "operationId": "changeNodeLifecycle",
        "summary": "Suspend or terminate tenant on one node",
        "tags": [
          "Nodes"
        ],
        "description": "Call both nodes. Resume through /control/provision with a higher version. Termination removes containers but preserves data and backups; no revival under the same ID.",
        "responses": {
          "200": {
            "description": "Inspect the response body state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeLifecycleResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON; most other node validation failures return 503",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or timestamp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource or route not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Node validation or operational failure: operation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "Timestamp": [],
            "Signature": []
          }
        ],
        "x-hmac-secret": "NODE_CONTROL_SECRET",
        "servers": [
          {
            "url": "https://node-us.yourdomain.com",
            "description": "US node; replace hostname"
          },
          {
            "url": "https://node-de.yourdomain.com",
            "description": "Germany node; replace hostname"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeLifecycleInput"
              },
              "example": {
                "id": "t-111111111111111111111111",
                "fingerprint": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                "lifecycleVersion": 1,
                "action": "suspended"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Timestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Spartan-Timestamp",
        "description": "Decimal Unix milliseconds; calculate with signature for each request."
      },
      "Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Spartan-Signature",
        "description": "Lowercase HMAC-SHA256 of timestamp\\nMETHOD\\npathname\\nrawBody with the operation-specific shared secret."
      },
      "OriginSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Spartan-Origin",
        "description": "Literal ORIGIN_SECRET, for internal health and ingress requests only."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": true
      },
      "InitialAdmin": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Nonblank; no control characters."
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "writeOnly": true,
            "description": "No control characters. Always creates a superadmin; do not supply role."
          }
        },
        "required": [
          "displayName",
          "email",
          "password"
        ],
        "additionalProperties": false
      },
...
      "NodeProvisionInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^t-[a-f0-9]{24}$"
          },
          "primary": {
            "type": "string",
            "enum": [
              "us",
              "de"
            ]
          },
          "appKey": {
            "type": "string",
            "pattern": "^base64:[A-Za-z0-9+/]{43}=$",
            "writeOnly": true
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Must equal https://{id}.{BASE_DOMAIN}."
          },
          "fingerprint": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "lifecycleVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "default": 0
          },
          "initialAdmin": {
            "$ref": "#/components/schemas/InitialAdmin"
          }
        },
        "required": [
          "id",
          "primary",
          "appKey",
          "url",
          "fingerprint"
        ],
        "additionalProperties": true
      },
      "NodeProvisionResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^t-[a-f0-9]{24}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "provisioning"
            ]
          },
          "role": {
            "type": "string",
            "enum": [
              "primary",
              "secondary"
            ]
          }
        },
        "required": [
          "id",
          "status",
          "role"
        ],
        "additionalProperties": false
      },
      "NodeLifecycleInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^t-[a-f0-9]{24}$"
          },
          "fingerprint": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "lifecycleVersion": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "action": {
            "type": "string",
            "enum": [
              "suspended",
              "terminated"
            ]
          }
        },
        "required": [
          "id",
          "fingerprint",
          "lifecycleVersion",
          "action"
        ],
        "additionalProperties": true
      },
      "NodeLifecycleResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^t-[a-f0-9]{24}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "suspended",
              "terminated"
            ]
          },
          "lifecycleVersion": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "status",
          "lifecycleVersion"
        ],
        "additionalProperties": false
      },
...
    }
  }
}